add: classroom domain
